Split PHP-related sources in smaller files
