+++ reporter/sources/__init__.py
@@ -5,5 +5,5 @@ Export classes than will be used by "bin" scripts and tests
 from common import Source
 
 from caching import NotCachedWikiaApiResponsesSource
-from php import PHPErrorsSource, DBQueryErrorsSource, DBQueryNoLimitSource
+from reporter.sources.php import PHPErrorsSource, DBQueryErrorsSource, DBQueryNoLimitSource
 from pt_kill import KilledDatabaseQueriesSource
