+++ reporter/sources/php/__init__.py
@@ -1,0 +1,3 @@
+# expose all PHP-related sources
+from .db import DBQueryErrorsSource, DBQueryNoLimitSource
+from .errors import PHPErrorsSource
